Добавление проверки роли пользователя, ревью и изменение статуса

+++ src/main/java/ru/develop/schedule/application/impl/SprintServiceImpl.java
@@ -4,9 +4,16 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import ru.develop.schedule.application.services.ProjectPersonService;
+import ru.develop.schedule.domain.ProjectPerson;
+import ru.develop.schedule.domain.ProjectPersonId;
+import ru.develop.schedule.domain.enums.Role;
+import ru.develop.schedule.extern.exceptions.NoPermissionException;
 import ru.develop.schedule.extern.repositories.SprintRepository;
 import ru.develop.schedule.domain.Sprint;
 import ru.develop.schedule.application.services.SprintService;
+
+import java.util.Set;
 
 @Service
 @RequiredArgsConstructor
@@ -14,6 +21,8 @@ import ru.develop.schedule.application.services.SprintService;
 public class SprintServiceImpl implements SprintService {
 
     private final SprintRepository sprintRepository;
+
+    private final ProjectPersonService projectPersonService;
 
     @Transactional(readOnly = true)
     @Override
@@ -27,15 +36,19 @@ public class SprintServiceImpl implements SprintService {
 
     @Transactional
     @Override
-    public void createSprint(Sprint sprint) {
+    public void createSprint(Sprint sprint, Long projectId, Long personId) throws NoPermissionException {
+        projectPersonService.checkPermission(Set.of(Role.ROLE_ADMIN, Role.ROLE_SUPERVISOR, Role.ROLE_STUDENT), projectId, personId);
+
         sprintRepository.save(sprint);
         log.info("Sprint created");
     }
 
     @Transactional
     @Override
-    public void deleteSprint(Long sprintId) {
+    public void deleteSprint(Long sprintId, Long projectId, Long personId) throws NoPermissionException {
+        projectPersonService.checkPermission(Set.of(Role.ROLE_ADMIN, Role.ROLE_SUPERVISOR), projectId, personId);
         sprintRepository.deleteById(sprintId);
         log.info("Sprint with id {} deleted", sprintId);
     }
+
 }
